Linked repl to library source code
Now `cabal run` is working.

This is now a very nice little skeleton for a language experiment. It
has a repl, the boilerplate is together. The lexer and parser can now be
developed in a pure manner, taking a string as input (to `interpret`)
and producing a string as output.

Next I will add a parsec lexer and parser. I shall endeavor to separate
the language specific functionallity from the general core.